# swagger.yml
openapi: 3.0.0
# basePath: /api
servers:
  - url: https://localhost:5000/
info:
  title: Edge Manager API
  version: '1.2'
# Paths supported by the server application
paths:
  /users:
    get:
      tags:
        - User
      operationId: app.get_all_users
      summary: Get All users
      parameters:
        - in: header
          name: x-access-token
          schema:
            type: string
            format: token
          required: true
      responses:
        200:
          description: "All users"
  /user:
    post:
      tags:
        - User
      operationId: app.post_user
      summary: Create/update user
      requestBody:
        description: User credentials
        required: true
        content:
          application/json:
            schema:
              x-body-name: credentials
              type: object
              properties:
                name:
                  example: username
                  type: string
                password:
                  example: test_pass
                  type: string
      responses:
        201:
          description: "Successfully created/updated user"
          content:
            application/text:
              schema:
                type: string
  /user/{id}:
    delete:
      tags:
        - User
      operationId: app.delete_user
      summary: Delete user by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        204:
          description: "Deleted"
        404:
          description: "Id not found"
  /token:
    post:
      tags:
        - User
      operationId: app.get_token
      summary: Get API authorization token
      requestBody:
        description: User credentials
        required: true
        content:
          application/json:
            schema:
              x-body-name: credentials
              type: object
              properties:
                name:
                  example: mitko
                  type: string
                password:
                  example: 111
                  type: string
      responses:
        200:
          description: "Successfully got token"
          content:
            application/text:
              schema:
                type: string